+++ CHANGELOG.rst
@@ -4,5 +4,6 @@ CHANGELOG
 Unreleased
 ----------
 
+- WOM-504: Add error callbacks
 - WOM-449: Implement pymodbus interface
 - WOM-440: Implement simulation